http://bugzilla.zimbra.com/show_bug.cgi?id=13831 bug 13831
When catalina.out is > 2GB, fopen(3) fails (launcher opens
catalina.out to fchown it to the zimbra user instead of root).
This results in stdout/stderr going to launcher's stdout/stderr
instead of catalina.out.

Fixed this by using open(2) with O_LARGEFILE and adding #define
_GNU_SOURCE (latter gives you _LARGEFILE64_SOURCE support).

I can not, however, reproduce a "crash" when catalina.out is >
2GB on either RHEL4 or Ubuntu 6.06.1 LTS.  Will assing to Matt
for repro.

Copied from Perforce
 Change: 51177