Update the default value of JAVA_BINARY
The location of the Zimbra Java binary changed from /opt/zimbra/java/bin/java
to /opt/zimbra/common/bin/java.  So when building zmmailboxdmgr you have to
provide the new path to the Makefile to make it work.  There is an override to
do so in zm-build but since the Java binary will probably stay where it is
now, this patch changes the default path as well.